requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
environs==14.1.1
gunicorn==23.0.0
marshmallow==4.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
style==1.1.0
update==0.0.1
whitenoise==6.9.0
